requirements.txt
Django==2.1.2
pymysql
mysqlclient
gunicorn==19.8.1
gevent==1.3.1
djangorestframework==3.8.2
django-simple-captcha
django-filter==2.0.0
oauthlib==2.1.0
six==1.12.0
typing==3.6.6
chardet==3.0.4
django-oauth-toolkit==1.2.0
celery==4.2.1
requests
colorlog==3.1.4
django-redis==4.10.0
tldextract
django-aliyun-oss2
django-cors-headers
rest-condition==1.0.3
